backend/requirements.txt
fastapi==0.136.1
uvicorn[standard]==0.46.0
sqlalchemy==2.0.49
psycopg[binary]==3.3.4
pydantic==2.13.3
pydantic-settings==2.14.0
python-dotenv==1.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
email-validator==2.3.0
python-multipart==0.0.21
requests==2.32.5
redis==7.1.0
httpx==0.28.1
pytest==9.0.2
pytest-asyncio==1.3.0
alembic==1.17.2
joblib==1.5.2
numpy==2.3.5
scikit-learn==1.8.0
